lax
Trait Solve_

pub trait Solve_: Scalar + Sized {
    // Required methods
    fn lu(l: MatrixLayout, a: &mut [Self]) -> Result<Pivot>;
    fn inv(l: MatrixLayout, a: &mut [Self], p: &Pivot) -> Result<()>;
    fn solve(
        l: MatrixLayout,
        t: Transpose,
        a: &[Self],
        p: &Pivot,
        b: &mut [Self],
    ) -> Result<()>;
}

Required Methods

fn lu(l: MatrixLayout, a: &mut [Self]) -> Result<Pivot>

Computes the LU factorization of a general m x n matrix a using partial pivoting with row interchanges.

$ PA = LU $

Error

LapackComputationalFailure { return_code } when the matrix is singular

Division by zero will occur if it is used to solve a system of equations because U[(return_code-1, return_code-1)] is exactly zero.

fn inv(l: MatrixLayout, a: &mut [Self], p: &Pivot) -> Result<()>

fn solve( l: MatrixLayout, t: Transpose, a: &[Self], p: &Pivot, b: &mut [Self], ) -> Result<()>

Object Safety

This trait is not object safe.